requirements.txt
streamlit
pandas
numpy
onnxruntime
scikit-learn
pillow
pytesseract
transformers
